refactor: 로띠애니메이션 컴포넌트 위치 변경 (FE-128)
